requirements.txt
asgiref==3.9.1
dj-rest-auth==7.0.1
Django==5.2.4
django-allauth==65.10.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.9.0
